renamed genre func file
