+++ packages/react-frontend/src/utils/getGenres.js
